requirements.txt
torch
torchvision
torchmetrics
Pillow
opencv-python
scikit-image
numpy
matplotlib
tqdm
